добавление 4 UI компонентов

+++ vite.config.ts
@@ -5,14 +5,17 @@ import dts from 'vite-plugin-dts';
 import { viteStaticCopy } from 'vite-plugin-static-copy';
 import vue from '@vitejs/plugin-vue';
 import svgLoader from 'vite-svg-loader';
+const componentDirs = fs.readdirSync('./src/components').filter((dir) => 
+  fs.statSync(path.join('./src/components', dir)).isDirectory() && dir.startsWith('UI')
+);
 
-const files = fs.readdirSync('./src/components').filter((file) => file.includes('Ui'));
-
-const components = files.reduce<{ [key: string]: string }>((obj, component) => {
-  obj[component.split('.')[0]] = `src/components/${component}/${component}.vue`;
-
+const components = componentDirs.reduce<{ [key: string]: string }>((obj, dir) => {
+  const componentName = dir;
+  obj[componentName] = `src/components/${dir}/${dir}.vue`;
   return obj;
 }, {});
+
+components['index'] = 'src/components/index.ts';
 
 export default defineConfig({
   build: {
@@ -29,7 +32,13 @@ export default defineConfig({
       output: {
         globals: { vue: 'Vue' },
         entryFileNames: `[name]/[name].js`,
-        assetFileNames: `[name]/[name].[ext]`,
+        chunkFileNames: 'chunks/[name]-[hash].js',
+        assetFileNames: (assetInfo) => {
+          if (assetInfo.name?.endsWith('.css')) {
+            return 'styles/[name][extname]';
+          }
+          return 'assets/[name][extname]';
+        },
       },
     },
   },
@@ -41,22 +50,23 @@ export default defineConfig({
       cleanVueFileName: true,
       entryRoot: './src/components',
       exclude: ['**/*.stories.ts', '**/*.css'],
+      outDir: 'dist',
     }),
     {
       name: 'add-css-link',
       apply: 'build',
-
       writeBundle(option, bundle) {
         const cssFiles = Object.keys(bundle)
-          .filter((file) => file.endsWith('.css') && !file.includes('-'))
+          .filter((file) => file.endsWith('.css'))
           .map((file) => file.replace('.css', ''));
 
         for (const file of cssFiles) {
           const filePath = path.resolve('', 'dist', `${file}.js`);
-          const cssImport = `import "./${file.split('/')[0]}.css";`;
-          const data = fs.readFileSync(filePath, { encoding: 'utf8' });
-
-          fs.writeFileSync(filePath, `${cssImport}\n${data}`);
+          if (fs.existsSync(filePath)) {
+            const cssImport = `import "./${file}.css";`;
+            const data = fs.readFileSync(filePath, { encoding: 'utf8' });
+            fs.writeFileSync(filePath, `${cssImport}\n${data}`);
+          }
         }
       },
     },
@@ -64,16 +74,8 @@ export default defineConfig({
     viteStaticCopy({
       targets: [
         {
-          src: 'src/components/index.ts',
-          dest: '',
-          rename: 'index.js',
-          transform: (contents) => contents.toString().replace(/.(vue|ts)/g, '.js'),
-        },
-        {
-          src: 'src/components/index.ts',
-          dest: '',
-          rename: 'index.d.ts',
-          transform: (contents) => contents.toString().replace(/.(vue|ts)/g, ''),
+          src: 'src/styles/*.scss',
+          dest: 'styles',
         },
       ],
     }),
